Moves Catalogue from package
