+++ src/main/java/com/codurance/allaboard/web/controllers/learningPath/LearningPathController.java
@@ -1,6 +1,6 @@
 package com.codurance.allaboard.web.controllers.learningPath;
 
-import com.codurance.allaboard.core.model.catalogue.Catalogue;
+import com.codurance.allaboard.web.views.Catalogue;
 import com.codurance.allaboard.core.model.catalogue.LearningPaths;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
