+++ src/main/java/com/codurance/allaboard/web/views/Catalogue.java
@@ -1,12 +1,13 @@
-package com.codurance.allaboard.core.model.catalogue;
+package com.codurance.allaboard.web.views;
 
+import com.codurance.allaboard.core.model.catalogue.LearningPath;
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 
 public class Catalogue implements Serializable {
 
-    private List<LearningPath> learningPaths;
+    private final List<LearningPath> learningPaths;
 
     public Catalogue(Iterable<LearningPath> learningPathsQuerySet) {
         this.learningPaths = new ArrayList<>();
